Fix hasher to use templates for any glm::vec<3, T>

+++ src/common/universe/Universe.cpp
@@ -1,12 +1,12 @@
 #include "Universe.h"
 
 namespace voxel_game::universe {
-	PhysicsObject * Universe::getPhysicsObject(const uint32_t id) {
+	PhysicsObject * Universe::getPhysicsObject(const uint64_t id) {
 		const auto it = mObjects.find(id);
 		return it == mObjects.end()? nullptr: it->second.get();
 	}
 
-	std::vector<uint32_t> Universe::getPhysicsObjectsInSector(const glm::i64vec3 &pos) {
+	std::vector<uint64_t> Universe::getPhysicsObjectsInSector(const glm::i64vec3 &pos) {
 		const auto it = mSectorPhysicsObjects.find(pos);
 		return it == mSectorPhysicsObjects.end()? std::vector<uint32_t>{}: it->second;
 	}
@@ -17,10 +17,10 @@ namespace voxel_game::universe {
 		mSectorPhysicsObjects[object->getSectorPos()].push_back(id);
 	}
 
-	void Universe::removePhysicsObject(const uint32_t id) {
+	void Universe::removePhysicsObject(const uint64_t id) {
 		const PhysicsObject *object = getPhysicsObject(id);
 		mObjects.erase(id);
-		std::vector<uint32_t> &physicsObjects = mSectorPhysicsObjects[object->getSectorPos()];
+		std::vector<uint64_t> &physicsObjects = mSectorPhysicsObjects[object->getSectorPos()];
 		for (uint32_t i = 0; i < physicsObjects.size(); i++) {
 			if (physicsObjects[i] == id) {
 				physicsObjects.erase(physicsObjects.begin() + i);
